1_core_orchestrator/backend/pyproject.toml
[project]
name = "deer-flow"
version = "0.1.0"
description = "LangGraph-based AI agent system with sandbox execution capabilities"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "deerflow-harness",
    "chardet>=5,<6",
    "fastapi>=0.115.0",
    "httpx>=0.28.0",
    "python-multipart>=0.0.20",
    "sse-starlette>=2.1.0",
    "uvicorn[standard]>=0.34.0",
    "lark-oapi>=1.4.0",
    "slack-sdk>=3.33.0",
    "python-telegram-bot>=21.0",
    "langgraph-sdk>=0.1.51",
    "markdown-to-mrkdwn>=0.3.1",
    "ruamel-yaml>=0.19.1",
    "pillow>=12.1.1",
    "opencv-python-headless>=4.13.0.92",
    "scikit-learn>=1.7.0",
    "tenacity>=9.1.2",
    "transformers>=5.3.0",
    "torch==2.6.0",
    "torchvision==0.21.0",
    "ultralytics>=8.4.30",
    "segment-anything",
    "torchxrayvision",
    "loguru>=0.7.3",
    "asgi-correlation-id>=4.3.4",
]

[dependency-groups]
dev = ["pytest>=8.0.0", "ruff>=0.14.11"]

[tool.uv.workspace]
members = ["packages/harness"]

[tool.uv.sources]
deerflow-harness = { workspace = true }
